requirements-dev.txt
-e .
-e ./docs/mkdocs_include_markdown_plugin
pytest
mkdocs==1.0.4
pymdown-extensions
pip-tools
requests
click
mypy
pyyaml
typing-extensions
